+++ RELEASE_NOTES.md
@@ -1,3 +1,15 @@
+### 0.2.1  - 2020-04-17
+* Fix #6: Show played card
+* Show active bonus on players
+* Fix #16: make watchdog visible
+* Fix #15: make highvoltage visible
+
+
+### 0.2.0  - 2020-04-17
+* Use flex boxes for players dashboard
+* smaller font for smaller screens
+* Hide card text for cards that need to select target on board
+
 ### 0.1.8  - 2020-04-16
 * Fix #22: Bad move display when selecting first crossroad
 * Fix #24: Skip turn if no move is possible